+++ src/platform_impl/ios/uikit/application.rs
@@ -1,4 +1,4 @@
-use objc2::foundation::{CGRect, MainThreadMarker, NSArray, NSObject};
+use icrate::Foundation::{CGRect, MainThreadMarker, NSArray, NSObject};
 use objc2::rc::{Id, Shared};
 use objc2::{extern_class, extern_methods, msg_send_id, ClassType};
 
@@ -20,11 +20,10 @@ extern_methods!(
             unsafe { msg_send_id![Self::class(), sharedApplication] }
         }
 
-        pub fn windows(&self) -> Id<NSArray<UIWindow, Shared>, Shared> {
-            unsafe { msg_send_id![self, windows] }
-        }
+        #[method_id(windows)]
+        pub fn windows(&self) -> Id<NSArray<UIWindow, Shared>, Shared>;
 
-        #[sel(statusBarFrame)]
+        #[method(statusBarFrame)]
         pub fn statusBarFrame(&self) -> CGRect;
     }
 );