+++ src/platform_impl/ios/uikit/window.rs
@@ -1,6 +1,6 @@
-use objc2::foundation::NSObject;
 use objc2::rc::{Id, Shared};
-use objc2::{extern_class, extern_methods, msg_send_id, ClassType};
+use objc2::runtime::NSObject;
+use objc2::{extern_class, extern_methods, ClassType};
 
 use super::{UIResponder, UIScreen, UIView};
 
@@ -16,20 +16,19 @@ extern_class!(
 
 extern_methods!(
     unsafe impl UIWindow {
-        pub fn screen(&self) -> Id<UIScreen, Shared> {
-            unsafe { msg_send_id![self, screen] }
-        }
+        #[method_id(screen)]
+        pub fn screen(&self) -> Id<UIScreen, Shared>;
 
-        #[sel(setScreen:)]
+        #[method(setScreen:)]
         pub fn setScreen(&self, screen: &UIScreen);
 
-        #[sel(setHidden:)]
+        #[method(setHidden:)]
         pub fn setHidden(&self, flag: bool);
 
-        #[sel(makeKeyAndVisible)]
+        #[method(makeKeyAndVisible)]
         pub fn makeKeyAndVisible(&self);
 
-        #[sel(isKeyWindow)]
+        #[method(isKeyWindow)]
         pub fn isKeyWindow(&self) -> bool;
     }
 );